feat: move TopHatFilter and GaussianFilter classes to filters.py
